small fixes

+++ src/app/components/buttons/addToWishlistButton.tsx
@@ -5,10 +5,21 @@ import {AddToWishlistButtonProps} from "@/types";
 
 const AddToWishlistButton: React.FC<AddToWishlistButtonProps> = ({ item, itemType }) => {
     const addToWishlist = () => {
-        // Check if the item is already in the wishlist
+        // Retrieve the wishlist from local storage
         const wishlist = JSON.parse(localStorage.getItem('wishlist') || '[]');
-        // Check if the item is already in the wishlist
+        // Create an item with type
         const itemWithType = { ...item, type: itemType };
+        // Check if the item is already in the wishlist
+        const itemExist = wishlist.some((wishlistItem: any) =>
+            wishlistItem.id === itemWithType.id && wishlistItem.type === itemWithType.type
+        );
+
+        if (itemExist) {
+            alert('Already in wishlist!');
+            return;
+        }
+
+        // Add the item to the wishlist
         wishlist.push(itemWithType);
         // Save the updated wishlist to local storage
         localStorage.setItem('wishlist', JSON.stringify(wishlist));
